Fix EHTML parser
The current code assumes {X1, X2, X3} when is_list(X3), which is not always the case.

Running a v1 discovery against ssocircle.com returns the following:

<html><head>
<link rel="openid.server" href="http://idp.ssocircle.com/sso/idservice">
</head><body>

(note the single <link> element)

This results in an EHTML of the form {_, _, {link, ...} which crashes the code when we try
to apply {link, ..} as a parameter to lists:foldl().  This adds code to handle the non-list
types.

Signed-off-by: Gregory Haskins <ghaskins@novell.com>